+++ CHANGELOG.md
@@ -43,6 +43,7 @@ Please make sure to add your changes to the appropriate categories:
 - Added `TimedPeakHold<T>` dt-aware decay
 - Added `TimedEnvelope<T>` dt-aware follower
 - Added `AnalogBiquad<T>` s-domain section
+- Added `butterworth_lowpass` normalized prototype
 
 ### Changed